Move `listeners.go` under `/service`.
